Version: 1.6.5 [2009-10-30] o ROBUSTIFICATION: Lowered the risk for save() of Object to leave an   imcomplete file due to say power failures etc.  This is done by   first writing to a temporary file, which is then renamed.  If the   temporary file already exists, an exception is thrown.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/r-dots/pkg/R.oo@303 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ R/050.Object.R
@@ -782,7 +782,7 @@ setMethodS3("save", "Object", function(this, file=NULL, path=NULL, compress=TRUE
     # Write to a temporary file
     file <- pathnameT;
     on.exit({
-      if (file.exists(pathnameT)) {
+      if (!is.null(pathnameT) && file.exists(pathnameT)) {
         file.remove(pathnameT);
       }
     }, add=TRUE);
@@ -800,6 +800,7 @@ setMethodS3("save", "Object", function(this, file=NULL, path=NULL, compress=TRUE
     if (!file.exists(pathname) || file.exists(pathnameT)) {
       throw("Failed to rename temporary file: ", pathnameT, " -> ", pathname);
     }
+    pathnameT <- NULL;
     file <- pathname;
   }
